Reorganize javascript
