+++ app/views/shared/_feedback.html.erb
@@ -20,17 +20,17 @@
         </p>
         <div class="form-group">
           <label>Nome (opcional)</label>
-          <%= text_field_tag 'name', nil, class: 'form-control' %>
+          <%= text_field_tag 'name_for_feedback', nil, class: 'form-control' %>
         </div>
 
         <div class="form-group">
           <label>Email (opcional)</label>
-          <%= text_field_tag 'email', nil, class: 'form-control' %>
+          <%= text_field_tag 'email_for_feedback', nil, class: 'form-control' %>
         </div>
 
         <div class="form-group">
           <label>Mensagem</label>
-          <%= text_area_tag 'message', nil, class: 'form-control' %>
+          <%= text_area_tag 'message_for_feedback', nil, class: 'form-control' %>
         </div>
       </div>
       <div class="modal-footer">
