+++ config/importmap.rb
@@ -1,6 +1,9 @@
 # Pin npm packages by running ./bin/importmap
 
 pin "application", preload: true
+
+pin_all_from "app/javascript/lib", under: "/lib"
+
 pin "@hotwired/turbo-rails", to: "turbo.min.js", preload: true
 pin "@hotwired/stimulus", to: "stimulus.min.js", preload: true
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js", preload: true
